Move communes, departements and regions documentation to /docs route
