+++ pages/docs/communes.js
@@ -1,9 +1,9 @@
 import FaMapSigns from 'react-icons/lib/fa/map-signs'
 
-import Head from '../components/head'
-import Page from '../layouts/main'
+import Head from '../../components/head'
+import Page from '../../layouts/main'
 
-import Communes from '../components/communes/communes'
+import Communes from '../../components/communes/communes'
 
 const title = 'Communes'
 const description = 'Interrogez les référentiels géographiques plus facilement.'
