+++ pages/docs/regions.js
@@ -1,9 +1,9 @@
 import FaMapSigns from 'react-icons/lib/fa/map-signs'
 
-import Head from '../components/head'
-import Page from '../layouts/main'
+import Head from '../../components/head'
+import Page from '../../layouts/main'
 
-import Regions from '../components/regions/regions'
+import Regions from '../../components/regions/regions'
 
 const title = 'Régions'
 const description = 'Interrogez les référentiels géographiques plus facilement.'
